# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompareOut": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunOut"
            },
            "title": "Runs",
            "type": "array"
          },
          "timeseries": {
            "items": {
              "$ref": "#/components/schemas/TimeseriesOut"
            },
            "title": "Timeseries",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "timeseries"
        ],
        "title": "CompareOut",
        "type": "object"
      },
      "CompareRequest": {
        "properties": {
          "run_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 4,
            "minItems": 2,
            "title": "Run Ids",
            "type": "array"
          }
        },
        "required": [
          "run_ids"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExceedanceOut": {
        "properties": {
          "end_day": {
            "title": "End Day",
            "type": "integer"
          },
          "factors": {
            "items": {
              "type": "string"
            },
            "title": "Factors",
            "type": "array"
          },
          "limit_value": {
            "title": "Limit Value",
            "type": "number"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "peak_ratio": {
            "title": "Peak Ratio",
            "type": "number"
          },
          "peak_value": {
            "title": "Peak Value",
            "type": "number"
          },
          "start_day": {
            "title": "Start Day",
            "type": "integer"
          }
        },
        "required": [
          "metric",
          "start_day",
          "end_day",
          "peak_value",
          "limit_value",
          "peak_ratio",
          "factors"
        ],
        "title": "ExceedanceOut",
        "type": "object"
      },
...
      "ModelOut": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "name",
          "version",
          "description"
        ],
        "title": "ModelOut",
        "type": "object"
      },
...
      "RunOut": {
        "examples": [
          {
            "created_at": "2026-01-01T00:00:00",
            "decision_use_permitted": false,
            "exceedances": [],
            "horizon": "3_months",
            "id": 1,
            "kpis": {
              "compliance_pct": 100.0
            },
            "model_id": "effluent_v1",
            "model_version": "1.4.0",
            "plant_id": 1,
            "scenario_id": 1,
            "scenario_name": "Synthetic baseline",
            "status": "completed",
            "validation_status": "illustrative_unvalidated"
          }
        ],
        "properties": {
          "artifact_digest": {
            "default": "historical_unknown",
            "title": "Artifact Digest",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision_use_permitted": {
            "const": false,
            "default": false,
            "title": "Decision Use Permitted",
            "type": "boolean"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "exceedances": {
            "items": {
              "$ref": "#/components/schemas/ExceedanceOut"
            },
            "title": "Exceedances",
            "type": "array"
          },
          "horizon": {
            "title": "Horizon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "input_snapshot": {
            "additionalProperties": true,
            "title": "Input Snapshot",
            "type": "object"
          },
          "kpis": {
            "additionalProperties": true,
            "title": "Kpis",
            "type": "object"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "plant_id": {
            "title": "Plant Id",
            "type": "integer"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "integer"
          },
          "scenario_name": {
            "title": "Scenario Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validation_status": {
            "const": "illustrative_unvalidated",
            "default": "illustrative_unvalidated",
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "scenario_name",
          "plant_id",
          "horizon",
          "model_id",
          "model_version",
          "status",
          "kpis",
          "exceedances",
          "created_at"
        ],
        "title": "RunOut",
        "type": "object"
      },
      "RunRequest": {
        "properties": {
          "model_id": {
            "default": "effluent_v1",
            "title": "Model Id",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
...
      "TimeseriesOut": {
        "properties": {
          "compliance_limits": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Compliance Limits",
            "type": "object"
          },
          "dates": {
            "items": {
              "type": "string"
            },
            "title": "Dates",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "series": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Series",
            "type": "object"
          }
        },
        "required": [
          "run_id",
          "dates",
          "series",
          "compliance_limits"
        ],
        "title": "TimeseriesOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Monash WWTP Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/compare": {
      "post": {
        "operationId": "compare_runs_api_v1_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Compare Runs",
        "tags": [
          "simulations"
        ]
      }
    },
...
    "/api/v1/models": {
      "get": {
        "operationId": "get_models_api_v1_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelOut"
                  },
                  "title": "Response Get Models Api V1 Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Models",
        "tags": [
          "simulations"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/runs/{run_id}/replay": {
      "post": {
        "operationId": "replay_api_v1_runs__run_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replay",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/runs/{run_id}/timeseries": {
      "get": {
        "operationId": "get_run_timeseries_api_v1_runs__run_id__timeseries_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeseriesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Timeseries",
        "tags": [
          "simulations"
        ]
      }
    },
...
    "/api/v1/scenarios/{scenario_id}/run": {
      "post": {
        "operationId": "run_scenario_api_v1_scenarios__scenario_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Scenario",
        "tags": [
          "simulations"
        ]
      }
    }
  }
}